+++ client/src/test/unit/lib/helper.spec.ts
@@ -5,7 +5,6 @@ import {
   forceArray,
   getClassifiedAs,
   getIdentifiedByContent,
-  isSpecimen,
   transformDate,
   getContentByClassifiedAs,
   getDateContent,
@@ -142,26 +141,6 @@ describe('helper functions', () => {
     })
   })
 
-  describe('isSpecimen', () => {
-    it('returns true', () => {
-      const specimenType = [
-        config.aat.fossil,
-        config.aat.animalSpecimens,
-        config.aat.plantSpecimens,
-        config.aat.biologicalSpecimens,
-      ]
-      const specimen = isSpecimen(
-        specimenType[Math.floor(Math.random() * specimenType.length)],
-      )
-      expect(specimen).toBeTruthy()
-    })
-
-    it('returns false', () => {
-      const specimen = isSpecimen('not a specimen')
-      expect(specimen).toBeFalsy()
-    })
-  })
-
   describe('getIdentifiedByContent', () => {
     it('returns true', () => {
       const identifiedBy = [
@@ -259,17 +238,22 @@ describe('helper functions', () => {
       const date = transformDate('-1997-05-13T23:59:59Z')
       expect(date).toEqual('5/13/1998 BCE')
     })
+
+    it('returns transformed empty string if year is greater than or equal to 9999', () => {
+      const date = transformDate('9999-05-13T23:59:59Z')
+      expect(date).toEqual('')
+    })
+
+    it('returns transformed empty string if year is less than or equal to -9999', () => {
+      const date = transformDate('-9999-05-13T23:59:59Z')
+      expect(date).toEqual('')
+    })
   })
 
   describe('addOneToBceYear', () => {
     it('returns BCE year plus one', () => {
       const date = addOneToBceYear(1000)
       expect(date).toEqual(1001)
-    })
-
-    it('returns transformed dates BCE', () => {
-      const date = transformDate('-1997-05-13T23:59:59Z')
-      expect(date).toEqual('5/13/1998 BCE')
     })
   })